SNOW-2432713: Ensure that dummy_row_pos_mode passed to OrderedDataFrame.join() matches SnowflakeQueryCompiler._dummy_row_pos_mode

+++ src/snowflake/snowpark/modin/plugin/docstrings/series_utils.py
@@ -2642,7 +2642,7 @@ class CombinedDatetimelikeProperties:
         0    1
         1    2
         2    3
-        dtype: int64
+        dtype: int8
 
         For TimedeltaIndex:
 
@@ -2650,7 +2650,7 @@ class CombinedDatetimelikeProperties:
         >>> tdelta_idx
         TimedeltaIndex(['0 days 00:00:01', '0 days 00:00:02', '0 days 00:00:03'], dtype='timedelta64[ns]', freq=None)
         >>> tdelta_idx.seconds
-        Index([1, 2, 3], dtype='int64')
+        Index([1, 2, 3], dtype='int8')
         """
 
     @property
@@ -2702,7 +2702,7 @@ class CombinedDatetimelikeProperties:
         0    1
         1    2
         2    3
-        dtype: int64
+        dtype: int8
 
         For TimedeltaIndex:
 
@@ -2712,7 +2712,7 @@ class CombinedDatetimelikeProperties:
                         '0 days 00:00:00.000003'],
                        dtype='timedelta64[ns]', freq=None)
         >>> tdelta_idx.microseconds
-        Index([1, 2, 3], dtype='int64')
+        Index([1, 2, 3], dtype='int8')
         """
 
     @property
@@ -2734,7 +2734,7 @@ class CombinedDatetimelikeProperties:
         0    1
         1    2
         2    3
-        dtype: int64
+        dtype: int8
 
         For TimedeltaIndex:
 
@@ -2744,7 +2744,7 @@ class CombinedDatetimelikeProperties:
                         '0 days 00:00:00.000000003'],
                        dtype='timedelta64[ns]', freq=None)
         >>> tdelta_idx.nanoseconds
-        Index([1, 2, 3], dtype='int64')
+        Index([1, 2, 3], dtype='int8')
         """
 
     @property
